Move ApiResponse to shared crate, replace backend response.rs with re-export
- Define ApiResponse<T> in shared/lib.rs with Serialize/Deserialize only
  (no Responder impl, avoiding the orphan rule)
- Add compile-time trait assertions for ApiResponse<()> in shared/lib.rs
- Replace backend/src/response.rs with a pub use re-export from the shared
  crate, preserving the crate::response::ApiResponse import path used by
  error.rs and integration test helpers
- Update backend/api/mod.rs to import ApiResponse from shared directly and
  return HttpResponse::Ok().json(...) from health_check (since ApiResponse
  no longer implements Responder in the shared crate)
