+++ src/init/templates/backend/src/response.rs
@@ -1,51 +1,3 @@
 // src/response.rs
 
-use actix_web::HttpRequest;
-use actix_web::HttpResponse;
-use actix_web::Responder;
-use actix_web::body::BoxBody;
-use actix_web::http::header::ContentType;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ApiResponse<T> {
-    pub success: bool,
-    pub data: Option<T>,
-    pub error: Option<String>,
-}
-
-impl<T: Serialize> ApiResponse<T> {
-    pub fn success(data: T) -> Self {
-        Self {
-            success: true,
-            data: Some(data),
-            error: None,
-        }
-    }
-
-    pub fn error(message: &str) -> Self {
-        Self {
-            success: false,
-            data: None,
-            error: Some(message.to_string()),
-        }
-    }
-}
-
-impl<T: Serialize> Responder for ApiResponse<T> {
-    type Body = BoxBody;
-
-    fn respond_to(self, _req: &HttpRequest) -> HttpResponse<Self::Body> {
-        match serde_json::to_string(&self) {
-            Ok(body) => HttpResponse::Ok()
-                .content_type(ContentType::json())
-                .body(body),
-            Err(e) => {
-                tracing::error!("Failed to serialize API response: {:?}", e);
-                HttpResponse::InternalServerError()
-                    .content_type(ContentType::plaintext())
-                    .body("Internal Server Error")
-            }
-        }
-    }
-}
+pub use {{ crate_name }}_shared::ApiResponse;
